fix: update, refactor, and working unit test
