+++ .github/workflows/development.yml
@@ -21,12 +21,17 @@ jobs:
         aws-secret-access-key: ${{ secrets.MYSECRET }}
         aws-region: ap-southeast-2
           
-    - name: Copy to dev.craiggunson.com bucket.
+    - name: Build and Test.
       run: |
         npm install
+        npm run test
         npm run build
+
+    - name: Copy to dev.craiggunson.com bucket.
+      run: |
         cd build
         aws s3 sync . s3://dev.craiggunson.com --acl public-read --cache-control "max-age=60"  --delete \
         --exclude '.git*' \
         --exclude '.github*'
         
+
